+++ Runtimes/Supplemental/Runtime/CMakeLists.txt
@@ -128,6 +128,7 @@ if(${PROJECT_NAME}_ENABLE_BACKTRACING)
     Backtrace+Codable.swift
     BacktraceFormatter.swift
     BacktraceJSONFormatter.swift
+    BacktracerThreadLocals.swift
     Base64.swift
     ByteSwapping.swift
     CachingMemoryReader.swift
